mem_ammount_kb now is u64

+++ src/kernel/mem/pmm.c
@@ -30,7 +30,7 @@ extern uint8_t kernel_start;
 extern uint8_t kernel_end;
 
 uint32_t mem_ammount_b;
-uint32_t mem_ammount_kb;
+uint64_t mem_ammount_kb;
 uint8_t *mem_bitmap;
 uint8_t *mem_start;
 uint32_t bitmap_size;
